[fix] 프로필 수정 할 때, 기존의 id 타입을 Integer에서 Long으로 수정

+++ src/main/java/kr/codesqaud/cafe/domain/user/dto/UserUpdateDto.java
@@ -18,7 +18,7 @@ public class UserUpdateDto {
 		return name;
 	}
 
-	public User toEntity(Integer id) {
+	public User toEntity(Long id) {
 		User convertedUser = new User(name, password, email);
 		convertedUser.setId(id);
 		return convertedUser;
